pyproject.toml
[build-system]
requires = ["uv_build >= 0.9.11, <0.10.0"]
build-backend = "uv_build"

[project]
name = "grt-testing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "matplotlib>=3.10.7",
    "pandas>=2.3.3",
    "pyyaml>=6.0.3",
    "seaborn>=0.13.2",
]

[dependency-groups]
dev = [
    "ty>=0.0.1a31",
]
# dynamic = ["version"]
